+++ drizzle.config.ts
@@ -3,7 +3,7 @@ import { defineConfig } from "drizzle-kit";
 
 export default defineConfig({
     out: "./drizzle",
-    schema: "./src/libs/db/schema.ts",
+    schema: "./src/libs/db/schema",
     dialect: "postgresql",
     dbCredentials: {
         url: env.DRIZZLE_DATABASE_URL!,
