+++ src/libs/db/schema/invoices.ts
@@ -1,0 +1,7 @@
+import { pgTable, text } from "drizzle-orm/pg-core";
+
+export const invoices = pgTable("invoice", {
+    id: text("id")
+        .primaryKey()
+        .$defaultFn(() => crypto.randomUUID()),
+});
