+++ src/libs/db/schema/types.ts
@@ -1,0 +1,3 @@
+import { pgEnum } from "drizzle-orm/pg-core";
+
+export const roleEnum = pgEnum("roles", ["user", "staff"]);
