+++ src/libs/db/schema/users.ts
@@ -1,0 +1,14 @@
+import { pgEnum, pgTable, text, timestamp } from "drizzle-orm/pg-core";
+
+export const roleEnum = pgEnum("roles", ["user", "staff"]);
+export const users = pgTable("user", {
+    id: text("id")
+        .primaryKey()
+        .$defaultFn(() => crypto.randomUUID()),
+    name: text("name"),
+    email: text("email").unique(),
+    emailVerified: timestamp("emailVerified", { mode: "date" }),
+    image: text("image"),
+    password: text("password"),
+    role: roleEnum("role").default("user").notNull(),
+});
