+++ src/libs/db/seed/profileSeed.ts
@@ -3,7 +3,8 @@ import "dotenv/config";
 import { eq } from "drizzle-orm";
 
 import { db } from "..";
-import { profiles, users } from "../schema";
+import { profiles } from "../schema/profiles";
+import { users } from "../schema/users";
 
 async function seedProfiles() {
     // Fetch all users
