+++ src/libs/db/types/profile.ts
@@ -1,4 +1,5 @@
-import { InferSelectModel  } from 'drizzle-orm';
-import { profiles } from '../schema';
+import { InferSelectModel } from "drizzle-orm";
+
+import { profiles } from "../schema/profiles";
 
 export type Profile = InferSelectModel<typeof profiles>;
